renames module
