+++ feature/subscriber-attributes/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="com.revenuecat.purchases.subscriber_attributes">
+        package="com.revenuecat.purchases.subscriberattributes">
 </manifest>
