+++ feature/subscriber-attributes/src/main/java/com/revenuecat/purchases/subscriberattributes/factories.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.common.attributes
+package com.revenuecat.purchases.subscriberattributes
 
 import com.revenuecat.purchases.common.caching.SubscriberAttributeMap
 import com.revenuecat.purchases.common.caching.SubscriberAttributesPerAppUserIDMap
